Adding PhysicsObj update methods

+++ Source/ACE.Server/Physics/Animation/Sequence.cs
@@ -52,7 +52,7 @@ namespace ACE.Server.Physics.Animation
 
         }
 
-        public void Update(float quantum, AFrame offsetFrame)
+        public void Update(double quantum, AFrame offsetFrame)
         {
 
         }
